examples/servers/dad-jokes/function/requirements.txt
mcp-openapi-proxy

# For testing, this module is installed from local files.
# Uncomment this line to build using the module from PyPi
# run-mcp-servers-with-aws-lambda >= 0.0.1
